pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "langserve>=0.3.3",
    "langsmith>=0.7.12",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
    "watchdog>=6.0.0",
]
